pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "oresat-star-tracker"
description = "OreSat Star Tracker OLAF app"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "GPL-3.0"}
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Embedded Systems",
]
dependencies = [
    "imagecodecs",
    "numpy",
    "imageio",
    "opencv-python-headless==4.6.0.66",
    "oresat-configs",
    "oresat-olaf>=3.0.0",
    "tifffile",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "build >= 1.3.0",
    "mypy >= 1.18.2",
    "pytest >= 9.0.1",
    "pytest-cov >= 7.0.0",
    "pytest-mock >= 3.15.1",
    "ruff >= 0.14.0",
    "setuptools",
    "setuptools-scm",
]

[project.scripts]
oresat-star-tracker = "oresat_star_tracker.__main__:main"

[tool.setuptools.packages.find]
exclude = ["docs*", "tests*"]

[tool.setuptools.package-data]
"*" = ["*.html", "*.so", "*.h", "*.i", "Makefile"]

[tool.coverage.run]
source = ["oresat_star_tracker"]

[tool.pytest.ini_options]
addopts = "--cov=oresat_star_tracker --cov-report=term-missing"
python_files = "tests/test_*.py"

[tool.black]
line_length = 100

[tool.pylama]
format = "pylint"
skip = "*/.tox/*,*/.env/,*/.git/*,*/.github/*,*/build/*"
linters = "pycodestyle,pyflakes,pylint,mccabe,mypy,radon"
# C0103:    Arguments are not snake_case naming style or too short
# E203:     Whitespace before ':' (black does this)
# W0613:    Unused argument
# R0902:    Too many instance attributes
# R901:     Too complex
# R0913:    Too many arguments
# R0914:    Too many local variables
# W1514:    Using open wth explicitly specifying an encoding
# W0707:    Consider explicitly re-raising
ignore = "C0103,E203,W0613,R0902,R901,R0913,R0914,E1101,W1514,W0707"
max_line_length = 100

[[tool.pylama.files]]
path = "*/__init__.py"
# W0611:    Imported but unused
# C0114:    Missing module docstring
ignore = "W0611,C0114"

[[tool.pylama.files]]
path = "tests/*"
# W0212:    Access to a protected member of a class
ignore = "W0212"

[[tool.mypy.overrides]]
module = "canopen,lost,olaf,tifffile,oresat_configs,cv2"
ignore_missing_imports = true

[tool.pylint.main]
extension-pkg-whitelist = ["cv2"]

[tool.isort]
profile = "black"
line_length = 100

[tool.setuptools_scm]
write_to = "oresat_star_tracker/_version.py"
